css編集中

+++ db/migrate/20220619093730_create_workouts.rb
@@ -2,7 +2,7 @@ class CreateWorkouts < ActiveRecord::Migration[7.0]
   def change
     create_table :workouts do |t|
       t.integer :user_id
-      t.numeric :body_weight
+      t.float :body_weight
       t.string :part
       t.text :memo
       t.integer :sleep
